requirements.txt
flask==3.1.2
flask_sqlalchemy==3.0.3
psycopg[binary]
python-dotenv==1.0.0